Added basic cart curl structure
the function still just fetches a regular item rather than a cartItem but I want to leave this here as a good checkpoint in case I mess something up later but don't want to change the keywords again.

Also made a slight change to the object being fetched in itemsRouter so that the item is made a constant with the name being searched later, rather than only fetching the name like before.