Added more functionality to cartRouter
Now that I have a baseline for the table in postgreSQL, I have a better ideal on how to actually get the data I want from these functions. Notably, despite the cartItems table only containing the id of the row, as well as the foreign keys of both the user and item associated, there's still a way to figure out the name of the item through the fetch item function. I should definitely export and require it later but this is easier for now.