Basic display for a user's cart
Made another subroute where the cart route, followed by a user's id, can show the items the related user has in their cart. There is no verification as to make sure if the user logged in is the same as the user who's cart is being viewed.

In the future, consider rewriting to make it so that the user id is grabbed from the logged in session rather than the url, after sufficient testing anyways to more quickly make sure the related functions and queries work.